WIP
The implementation example is mostly garbage; it was abandoned halfway
through in favour of rewriting/extending the specification.

The newspec.md contains a WIP state of the document. It is likely still
internally inconsistent, jumps around between various unexplained ideas,
and also may contain various phrases that just end in the middle because
I haven't finished writing things out.